Added a factorial.cpp example.
