+++ Makefile.am
@@ -1,6 +1,6 @@
 SUBDIRS = src tests examples
 
-ACLOCAL_AMFLAGS = -I m4
+
 
 docs:
 	@doxygen Doxyfile
