+++ examples/Makefile.am
@@ -8,7 +8,9 @@ noinst_PROGRAMS = \
 	selection_sort \
 	merge_sort \
 	progress \
-	producer_consumer
+	producer_consumer \
+	fibonacci \
+	factorial
 
 
 AM_CXXFLAGS = -g -Wall -I$(top_srcdir)/include -I/usr/lib/gcc/x86_64-pc-linux-gnu/9.1.0/include -fopenmp -pthread
@@ -42,3 +44,9 @@ progress_LDADD = $(top_builddir)/src/libtsal.la
 
 producer_consumer_SOURCES = producer_consumer.cpp
 producer_consumer_LDADD = $(top_builddir)/src/libtsal.la
+
+fibonacci_SOURCES = fibonacci.cpp
+fibonacci_LDADD = $(top_builddir)/src/libtsal.la
+
+factorial_SOURCES = factorial.cpp
+factorial_LDADD = $(top_builddir)/src/libtsal.la
